grpc_app/requirements.txt
grpcio
grpcio-tools
torch
torchvision
pillow
numpy<2
